requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
python-dotenv==1.0.0
tinify==1.6.0
Pillow==10.0.0
pymongo==4.5.0
APScheduler==3.10.1
validators==0.22.0
requests==2.31.0
aiohttp==3.8.4
aiofiles==23.1.0